requirements.txt
pygame>=2.5.0
python-chess>=1.9.4
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0